Refactor: test de login social
